split namco c169 roz chip from namcos2 by making it a device (nw)
some reorganization still needed (eg move init stuff to config)
