+++ src/mame/video/namcofl.cpp
@@ -69,7 +69,7 @@ uint32_t namcofl_state::screen_update_namcofl(screen_device &screen, bitmap_ind1
 
 	for( pri=0; pri<16; pri++ )
 	{
-		c169_roz_draw(screen, bitmap, cliprect, pri);
+		m_c169roz->draw(screen, bitmap, cliprect, pri);
 		if((pri&1)==0)
 			c123_tilemap_draw( screen, bitmap, cliprect, pri>>1 );
 		c355_obj_draw(screen, bitmap, cliprect, pri );
@@ -98,5 +98,5 @@ VIDEO_START_MEMBER(namcofl_state,namcofl)
 {
 	c123_tilemap_init(NAMCOFL_TILEGFX, memregion(NAMCOFL_TILEMASKREGION)->base(), namcos2_shared_state::c123_tilemap_delegate(&namcofl_state::TilemapCB, this));
 	c355_obj_init(NAMCOFL_SPRITEGFX,0x0,namcos2_shared_state::c355_obj_code2tile_delegate(&namcofl_state::FLobjcode2tile, this));
-	c169_roz_init(NAMCOFL_ROTGFX,NAMCOFL_ROTMASKREGION,namcos2_shared_state::c169_tilemap_delegate(&namcofl_state::RozCB, this));
+	m_c169roz->init(NAMCOFL_ROTGFX,"^" NAMCOFL_ROTMASKREGION,namco_c169roz_device::c169_tilemap_delegate(&namcofl_state::RozCB, this));
 }
