+++ src/mame/video/namconb1.cpp
@@ -94,7 +94,7 @@ void namconb1_state::video_update_common(screen_device &screen, bitmap_ind16 &bi
 	{
 		for( pri=0; pri<16; pri++ )
 		{
-			c169_roz_draw(screen, bitmap, cliprect, pri);
+			m_c169roz->draw(screen, bitmap, cliprect, pri);
 			if( (pri&1)==0 )
 			{
 				c123_tilemap_draw( screen, bitmap, cliprect, pri/2 );
@@ -154,8 +154,7 @@ WRITE32_MEMBER(namconb1_state::rozbank32_w)
 	uint32_t old_data = m_rozbank32[offset];
 	COMBINE_DATA(&m_rozbank32[offset]);
 	if (m_rozbank32[offset] != old_data)
-		for (auto & elem : m_c169_roz_tilemap)
-			elem->mark_all_dirty();
+		m_c169roz->mark_all_dirty();
 }
 
 uint32_t namconb1_state::screen_update_namconb2(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
@@ -201,7 +200,7 @@ VIDEO_START_MEMBER(namconb1_state,machbrkr)
 {
 	c123_tilemap_init(NAMCONB1_TILEGFX, memregion(NAMCONB1_TILEMASKREGION)->base(), namcos2_shared_state::c123_tilemap_delegate(&namconb1_state::NB2TilemapCB_machbrkr, this));
 	c355_obj_init(NAMCONB1_SPRITEGFX,0x0,namcos2_shared_state::c355_obj_code2tile_delegate(&namconb1_state::NB2objcode2tile_machbrkr, this));
-	c169_roz_init(NAMCONB1_ROTGFX,NAMCONB1_ROTMASKREGION,namcos2_shared_state::c169_tilemap_delegate(&namconb1_state::NB2RozCB_machbrkr, this));
+	m_c169roz->init(NAMCONB1_ROTGFX,"^" NAMCONB1_ROTMASKREGION,namco_c169roz_device::c169_tilemap_delegate(&namconb1_state::NB2RozCB_machbrkr, this));
 
 	save_item(NAME(m_tilemap_tile_bank));
 } /* machbrkr */
@@ -210,7 +209,7 @@ VIDEO_START_MEMBER(namconb1_state,outfxies)
 {
 	c123_tilemap_init(NAMCONB1_TILEGFX, memregion(NAMCONB1_TILEMASKREGION)->base(), namcos2_shared_state::c123_tilemap_delegate(&namconb1_state::NB2TilemapCB_outfxies, this));
 	c355_obj_init(NAMCONB1_SPRITEGFX,0x0,namcos2_shared_state::c355_obj_code2tile_delegate(&namconb1_state::NB2objcode2tile_outfxies, this));
-	c169_roz_init(NAMCONB1_ROTGFX,NAMCONB1_ROTMASKREGION,namcos2_shared_state::c169_tilemap_delegate(&namconb1_state::NB2RozCB_outfxies, this));
+	m_c169roz->init(NAMCONB1_ROTGFX,"^" NAMCONB1_ROTMASKREGION,namco_c169roz_device::c169_tilemap_delegate(&namconb1_state::NB2RozCB_outfxies, this));
 
 	save_item(NAME(m_tilemap_tile_bank));
 } /* outfxies */
