+++ src/mame/video/namcos2.cpp
@@ -417,9 +417,9 @@ void namcos2_state::video_start_luckywld()
 {
 	c123_tilemap_init(2,memregion("gfx4")->base(),namcos2_shared_state::c123_tilemap_delegate(&namcos2_state::TilemapCB, this));
 	c355_obj_init( 0, 0x0, namcos2_shared_state::c355_obj_code2tile_delegate() );
-	if( m_gametype==NAMCOS2_LUCKY_AND_WILD )
+	if( m_gametype==NAMCOS2_LUCKY_AND_WILD ) // suzuka8h is using the same machine config, is it the same PCB just without the ROZ populated, or should it be split?
 	{
-		c169_roz_init(1, "gfx5", namcos2_shared_state::c169_tilemap_delegate(&namcos2_state::RozCB_luckywld, this));
+		m_c169roz->init(1, "^gfx5", namco_c169roz_device::c169_tilemap_delegate(&namcos2_state::RozCB_luckywld, this));
 	}
 }
 
@@ -440,7 +440,7 @@ uint32_t namcos2_state::screen_update_luckywld(screen_device &screen, bitmap_ind
 		m_c45_road->draw(bitmap,clip,pri);
 		if( m_gametype==NAMCOS2_LUCKY_AND_WILD )
 		{
-			c169_roz_draw(screen, bitmap, clip, pri);
+			m_c169roz->draw(screen, bitmap, clip, pri);
 		}
 		c355_obj_draw(screen, bitmap, clip, pri );
 	}
@@ -483,7 +483,7 @@ void namcos2_state::RozCB_metlhawk(uint16_t code, int *tile, int *mask, int whic
 void namcos2_state::video_start_metlhawk()
 {
 	c123_tilemap_init(2,memregion("gfx4")->base(),namcos2_shared_state::c123_tilemap_delegate(&namcos2_state::TilemapCB, this));
-	c169_roz_init(1, "gfx5", namcos2_shared_state::c169_tilemap_delegate(&namcos2_state::RozCB_metlhawk, this));
+	m_c169roz->init(1, "^gfx5", namco_c169roz_device::c169_tilemap_delegate(&namcos2_state::RozCB_metlhawk, this));
 }
 
 uint32_t namcos2_state::screen_update_metlhawk(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
@@ -500,7 +500,7 @@ uint32_t namcos2_state::screen_update_metlhawk(screen_device &screen, bitmap_ind
 		{
 			c123_tilemap_draw( screen, bitmap, clip, pri/2 );
 		}
-		c169_roz_draw(screen, bitmap, clip, pri);
+		m_c169roz->draw(screen, bitmap, clip, pri);
 		draw_sprites_metalhawk(screen,bitmap,clip,pri );
 	}
 	return 0;
